pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "fast_priority"
version = "0.0.1"
description = "Simple and streigt forward priority queing gateway using fastapi"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.115.8",
  "httpx",
  "python-dotenv>=1.0.1",
  "rq>=2.1.0",
  "click",
  "uvicorn>=0.34.0"
]

[project.scripts]
run-fast-priority-queue = "fast_priority.cli:main"

[tool.uv]
default-groups = ["dev", "test"]

[dependency-groups]
test = [
  "pytest",
  "pytest-mock",
  "pytest-cov"
]
dev = [
  "uv",
  "ruff",
  "pre-commit",
  "bpython",
  "debugpy"
]

[tool.pytest.ini_options]
testpaths = ["tests/unit/"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "PD",
  "PT",
  "NPY",
  "N",
  "W",
  "RET",
  "ANN",
  "RUF",
  "SIM",
  "PERF",
]
unfixable = ["F841", "RUF015"]
ignore = [
  "ANN202",
  "ANN205",
  "ANN001",
  "ANN002",
  "ANN003",
  "ANN401",
  "RET502",
  "RET503",
  "RET504",
  "RET505",
  "RUF005",
  "RUF009",
  "E731",
]
